refactor(StandardDeck): add StandardDeckCard model and view
* Separate the model and the view
* Make the model `Hashable ` and `Identifiable` from just the Rank and
  Suit to avoid issue when rendering in SwiftUI views with updates
* Split the `StandardDeckCardView` with multiple views that uses
  `Content: View` to avoid confusions and improve readability and
  maintainability
